botões e controller
 commit da criação de todos os botões e todos funcionando a principio e acrescentando os controller para cada tela de cadastro

+++ src/bean/VccFornecedor.java
@@ -176,7 +176,7 @@ public class VccFornecedor  implements java.io.Serializable {
         return this.vccPreco;
     }
     
-    public void setVccPreco(int vccPreco) {
+    public void setVccPreco(double vccPreco) {
         this.vccPreco = vccPreco;
     }
