+++ src/bean/VccProdutos.java
@@ -103,7 +103,7 @@ public class VccProdutos  implements java.io.Serializable {
         return this.vccPreco;
     }
     
-    public void setVccPreco(int vccPreco) {
+    public void setVccPreco(double vccPreco) {
         this.vccPreco = vccPreco;
     }
